Don't delete graphs in virtual DOM
If they are in the virtual DOM but not in the actual DOM, they are
about to be added to the DOM, not removed.

This was a mistake in commit fc6c073e0d58fc243c3987d541b9f6356d7f26f7,
which despite the commit message affects all graphs, not just editor
graphs.