Use shared web workers for graphviz renderers to improve performance
With dedicated web workers, one worker was created for every
graph. Each web worker loaded its own (huge) WASM script. When editing
a post, for every key stroke all the graphviz renderers in the post
are removed and re-created. This caused severe performance problems.

With this change, the shared web worker is create only once and each
new graphviz renderer just creates a port to it when it is
created. This port is removed when the graphviz renderer is
destroyed.

Together with commit 6d281aff9265aad5e291031e897ae7980030a6ec this
should make a huge user experience improvement.