Fix editor preview scroll jumps when editor is scrolled
The collapsing of graphs that are not visible in the editor preview
that was introduced in commit ffafddf97b65f9907142054b4b5951603aeaa6e3
made synchronized scrolling between the editor and the editor preview
somewhat better, but makes the preview jump when graphs are hidden.

There is still a jump the first time a graph is rendered at the bottom
or top of the preview, but it's acceptable since it's a one-off. The
alternative is to render all graphs in the preview for every keystroke
and that's not acceptable.