Fix non-empty paragraphs are removed after DOT BBCode extraction
Removal of paragraphs that are empty after DOT BBCode extraction was
added in commit 3bd307aec8965d2d54724667c702df21f2eaa71e. It was
assumed that all paragraphs except the one with the graph would be
empty, but that's not always the case, e.g. when text precedes a DOT
BBcode block.

Fixed by adding a check that the paragraph really is empty or contains
just a single line break before removing it.