fix empty paragraphs left behind
This reverts the fix in 083273ce7c169a360769d0817747dbd133d15fa6 ("Fix
editor preview stops updating after editing") which stated:

    BBCode element parents that were themselves BBCode elements were
    removed twice.

This seems to have been incorrect:

  1. There doesn't seem to be any BBCode element parents that were
     themselves BBCode elements

  2. The fix incorrectly excluded elements to be considered for
     removal which did *not* fulfill this condition, which were all
     elements. Therefore no parent elements were removed and empty
     paragraphs were incorrectly kept.

The new fix implements a workaround which hides the empty paragraphs
instead of removing them, since that is what caused the problems in
https://github.com/magjac/discourse-d3-graphviz/issues/44.

This does not seem to have any visible effect, so no issue or
CHANGELOG entry has been written for this.

Also, from
https://meta.discourse.org/t/notfounderror-node-removechild-the-node-to-be-removed-is-not-a-child-of-this-node/154739/4:

------------------------------------------------------------------------
David Taylor david team Jun '20

The composer preview uses ember to render the cooked content. It looks
like that error is coming from the ember rendering engine, which is
probably confused by the DOM being changed by something else.

To avoid the issue, I suggest that you don’t remove() any nodes, and
instead hide them using display:none; or something similar.
------------------------------------------------------------------------

and from
https://meta.discourse.org/t/notfounderror-node-removechild-the-node-to-be-removed-is-not-a-child-of-this-node/154739/6:

------------------------------------------------------------------------
David Taylor david team Jun '20

The composer preview is an unusual combination of ember rendering and
custom ‘decorateCooked’ functions. It looks like this is a quirk of
that implementation. We don’t have any plans to change the way it
works at the moment, especially since the workaround is nice and
simple.

------------------------------------------------------------------------