fix flaky deleteCypressTestingTopic command
After deleting a topic and going back to the previous page (the topic
list), the just deleted topic could still be shown and an attempt to
delete it again would then fail. This happened about 50% of the
time. The fix is to reload the topic list page and thus flush the just
deleted topic.